Added directory html to stdhttp
This was because it is hard to download mp3s from mobile web browsers.
The browsers tend to want to play the mp3 rather than asking if you want
to save, having a webpage allows you to long press the link and then
choose save link. I also added 'Content-Length' to the header, to have a
proper progress bar and know how long a download will take.